possible synchronizer refactor solution
