+++ pkg/subscriber/github/git_subscriber_item.go
