+++ pkg/subscriber/namespace/deployable_reconciler.go
@@ -31,6 +31,7 @@ import (
 
 	dplv1alpha1 "github.com/open-cluster-management/multicloud-operators-deployable/pkg/apis/apps/v1"
 
+	kubesynchronizer "github.com/open-cluster-management/multicloud-operators-subscription/pkg/synchronizer/kubernetes"
 	"github.com/open-cluster-management/multicloud-operators-subscription/pkg/utils"
 )
 
@@ -152,9 +153,9 @@ func (r *DeployableReconciler) doSubscription() error {
 
 	hostkey := types.NamespacedName{Name: subitem.Subscription.Name, Namespace: subitem.Subscription.Namespace}
 	syncsource := deployablesyncsource + hostkey.String()
+
 	// subscribed k8s resource
-	kvalid := r.subscriber.synchronizer.CreateValiadtor(syncsource)
-	pkgMap := make(map[string]bool)
+	dplOrder := []kubesynchronizer.DplUnit{}
 
 	vsub := ""
 
@@ -167,7 +168,7 @@ func (r *DeployableReconciler) doSubscription() error {
 	for _, dpl := range dpllist.Items {
 		klog.V(5).Infof("Updating subscription %v, with Deployable %v  ", syncsource, hostkey)
 
-		dpltosync, validgvk, err := r.doSubscribeDeployable(subitem, dpl.DeepCopy(), versionMap, pkgMap)
+		dpltosync, validgvk, err := r.doSubscribeDeployable(subitem, dpl.DeepCopy(), versionMap, map[string]bool{})
 		if err != nil {
 			klog.V(3).Info("Skipping deployable", dpl.Name)
 
@@ -178,9 +179,13 @@ func (r *DeployableReconciler) doSubscription() error {
 			continue
 		}
 
-		klog.V(5).Info("Ready to register template:", hostkey, dpltosync, syncsource)
+		dplU := kubesynchronizer.DplUnit{
+			Dpl: dpltosync,
+			Gvk: *validgvk,
+		}
 
-		err = r.subscriber.synchronizer.RegisterTemplate(hostkey, dpltosync, syncsource)
+		dplOrder = append(dplOrder, dplU)
+		klog.V(5).Info("Ready to register template:", hostkey, dpltosync, syncsource)
 
 		if err != nil {
 			err = utils.SetInClusterPackageStatus(&(subitem.Subscription.Status), dpltosync.GetName(), err, nil)
@@ -189,8 +194,6 @@ func (r *DeployableReconciler) doSubscription() error {
 				klog.Info("error in setting in cluster package status :", err)
 			}
 
-			pkgMap[dpltosync.GetName()] = true
-
 			return errors.Wrap(err, "failed to update subscription status")
 		}
 
@@ -198,15 +201,11 @@ func (r *DeployableReconciler) doSubscription() error {
 			Name:      dpltosync.Name,
 			Namespace: dpltosync.Namespace,
 		}
-		kvalid.AddValidResource(*validgvk, hostkey, dplkey)
-
-		pkgMap[dplkey.Name] = true
 
 		klog.V(5).Info("Finished Register ", *validgvk, hostkey, dplkey, " with err:", err)
 	}
 
-	r.subscriber.synchronizer.ApplyValiadtor(kvalid)
-
+	retryerr = r.subscriber.synchronizer.AddTemplates(syncsource, hostkey, dplOrder)
 	return retryerr
 }
